requirements.txt
# LangChain Core
langchain==0.3.25



# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic
langchain-chroma

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
sentence-transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# Streamlit (optional)
streamlit

# Typing utilities
typing-extensions
